# spec/swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the Ory Kratos HTTP API documentation!",
    "title": "Ory Kratos",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/self-service/verification": {
      "post": {
        "description": "Use this endpoint to complete a verification flow. This endpoint\nbehaves differently for API and browser flows and has several states:\n\n`choose_method` expects `flow` (in the URL query) and `email` (in the body) to be sent\nand works with API- and Browser-initiated flows.\nFor API clients and Browser clients with HTTP Header `Accept: application/json` it either returns a HTTP 200 OK when the form is valid and HTTP 400 OK when the form is invalid\nand a HTTP 303 See Other redirect with a fresh verification flow if the flow was otherwise invalid (e.g. expired).\nFor Browser clients without HTTP Header `Accept` or with `Accept: text/*` it returns a HTTP 303 See Other redirect to the Verification UI URL with the Verification Flow ID appended.\n`sent_email` is the success state after `choose_method` when using the `link` method and allows the user to request another verification email. It\nworks for both API and Browser-initiated flows and returns the same responses as the flow in `choose_method` state.\n`passed_challenge` expects a `token` to be sent in the URL query and given the nature of the flow (\"sending a verification link\")\ndoes not have any API capabilities. The server responds with a HTTP 303 See Other redirect either to the Settings UI URL\n(if the link was valid) and instructs the user to update their password, or a redirect to the Verification UI URL with\na new Verification Flow ID which contains an error message that the verification link was invalid.\n\nMore information can be found at [Ory Kratos Email and Phone Verification Documentation](https://www.ory.sh/docs/kratos/selfservice/flows/verify-email-account-activation).",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Complete Verification Flow",
        "operationId": "submitSelfServiceVerificationFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Verification Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/verification?flow=abcde`).",
            "name": "flow",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Verification Token\n\nThe verification token which completes the verification request. If the token\nis invalid (e.g. expired) an error will be shown to the end-user.\n\nThis parameter is usually set in a link and not used by any direct API call.",
            "name": "token",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/submitSelfServiceVerificationFlowBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceVerificationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceVerificationFlow"
            }
          },
          "303": {
            "$ref": "#/responses/emptyResponse"
          },
          "400": {
            "description": "selfServiceVerificationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceVerificationFlow"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/verification/api": {
      "get": {
        "description": "This endpoint initiates a verification flow for API clients such as mobile devices, smart TVs, and so on.\n\nTo fetch an existing verification flow call `/self-service/verification/flows?flow=\u003cflow_id\u003e`.\n\nYou MUST NOT use this endpoint in client-side (Single Page Apps, ReactJS, AngularJS) nor server-side (Java Server\nPages, NodeJS, PHP, Golang, ...) browser applications. Using this endpoint in these applications will make\nyou vulnerable to a variety of CSRF attacks.\n\nThis endpoint MUST ONLY be used in scenarios such as native mobile apps (React Native, Objective C, Swift, Java, ...).\n\nMore information can be found at [Ory Kratos Email and Phone Verification Documentation](https://www.ory.sh/docs/kratos/selfservice/flows/verify-email-account-activation).",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Initialize Verification Flow for APIs, Services, Apps, ...",
        "operationId": "initializeSelfServiceVerificationFlowWithoutBrowser",
        "responses": {
          "200": {
            "description": "selfServiceVerificationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceVerificationFlow"
            }
          },
          "400": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/verification/browser": {
      "get": {
        "description": "This endpoint initializes a browser-based account verification flow. Once initialized, the browser will be redirected to\n`selfservice.flows.verification.ui_url` with the flow ID set as the query parameter `?flow=`.\n\nIf this endpoint is called via an AJAX request, the response contains the recovery flow without any redirects.\n\nThis endpoint is NOT INTENDED for API clients and only works with browsers (Chrome, Firefox, ...).\n\nMore information can be found at [Ory Kratos Email and Phone Verification Documentation](https://www.ory.sh/docs/kratos/selfservice/flows/verify-email-account-activation).",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Initialize Verification Flow for Browser Clients",
        "operationId": "initializeSelfServiceVerificationFlowForBrowsers",
        "parameters": [
          {
            "type": "string",
            "description": "The URL to return the browser to after the flow was completed.",
            "name": "return_to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceVerificationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceVerificationFlow"
            }
          },
          "303": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
    "/self-service/verification/flows": {
      "get": {
        "description": "This endpoint returns a verification flow's context with, for example, error details and other information.\n\nBrowser flows expect the anti-CSRF cookie to be included in the request's HTTP Cookie Header.\nFor AJAX requests you must ensure that cookies are included in the request or requests will fail.\n\nIf you use the browser-flow for server-side apps, the services need to run on a common top-level-domain\nand you need to forward the incoming HTTP Cookie header to this endpoint:\n\n```js\npseudo-code example\nrouter.get('/recovery', async function (req, res) {\nconst flow = await client.getSelfServiceVerificationFlow(req.header('cookie'), req.query['flow'])\n\nres.render('verification', flow)\n})\n\nMore information can be found at [Ory Kratos Email and Phone Verification Documentation](https://www.ory.sh/docs/kratos/selfservice/flows/verify-email-account-activation).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "v0alpha2"
        ],
        "summary": "Get Verification Flow",
        "operationId": "getSelfServiceVerificationFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Flow ID\n\nThe value for this parameter comes from `request` URL Query parameter sent to your\napplication (e.g. `/verification?flow=abcde`).",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "HTTP Cookies\n\nWhen using the SDK on the server side you must include the HTTP Cookie Header\noriginally sent to your HTTP handler here.",
            "name": "cookie",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "selfServiceVerificationFlow",
            "schema": {
              "$ref": "#/definitions/selfServiceVerificationFlow"
            }
          },
          "403": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "404": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          },
          "500": {
            "description": "jsonError",
            "schema": {
              "$ref": "#/definitions/jsonError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ID": {
      "type": "integer",
      "format": "int64"
    },
...
    "UUID": {"type": "string", "format": "uuid4"},
...
    "genericError": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "code": {
          "description": "The status code",
          "type": "integer",
          "format": "int64",
          "example": 404
        },
        "debug": {
          "description": "Debug information\n\nThis field is often not exposed to protect against leaking\nsensitive information.",
          "type": "string",
          "example": "SQL field \"foo\" is not a bool."
        },
        "details": {
          "description": "Further error details",
          "type": "object",
          "additionalProperties": true
        },
        "id": {
          "description": "The error ID\n\nUseful when trying to identify various errors in application logic.",
          "type": "string"
        },
        "message": {
          "description": "Error message\n\nThe error's message.",
          "type": "string",
          "example": "The resource could not be found"
        },
        "reason": {
          "description": "A human-readable reason for the error",
          "type": "string",
          "example": "User with ID 1234 does not exist."
        },
        "request": {
          "description": "The request ID\n\nThe request ID is often exposed internally in order to trace\nerrors across service architectures. This is often a UUID.",
          "type": "string",
          "example": "d7ef54b1-ec15-46e6-bccb-524b82c035e6"
        },
        "status": {
          "description": "The status description",
          "type": "string",
          "example": "Not Found"
        }
      }
    },
...
    "jsonError": {
      "description": "The standard Ory JSON API error format.",
      "type": "object",
      "title": "JSON API Error Response",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "$ref": "#/definitions/genericError"
        }
      }
    },
...
    "selfServiceFlowType": {
      "description": "The flow type can either be `api` or `browser`.",
      "type": "string",
      "title": "Type is the flow type."
    },
...
    "selfServiceVerificationFlow": {
      "description": "Used to verify an out-of-band communication\nchannel such as an email address or a phone number.\n\nFor more information head over to: https://www.ory.sh/docs/kratos/selfservice/flows/verify-email-account-activation",
      "type": "object",
      "title": "A Verification Flow",
      "required": [
        "id",
        "type",
        "ui",
        "state"
      ],
      "properties": {
        "active": {
          "description": "Active, if set, contains the registration method that is being used. It is initially\nnot set.",
          "type": "string"
        },
        "expires_at": {
          "description": "ExpiresAt is the time (UTC) when the request expires. If the user still wishes to verify the address,\na new request has to be initiated.",
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "issued_at": {
          "description": "IssuedAt is the time (UTC) when the request occurred.",
          "type": "string",
          "format": "date-time"
        },
        "request_url": {
          "description": "RequestURL is the initial URL that was requested from Ory Kratos. It can be used\nto forward information contained in the URL's path or query for example.",
          "type": "string"
        },
        "return_to": {
          "description": "ReturnTo contains the requested return_to URL.",
          "type": "string"
        },
        "state": {
          "$ref": "#/definitions/selfServiceVerificationFlowState"
        },
        "type": {
          "$ref": "#/definitions/selfServiceFlowType"
        },
        "ui": {
          "$ref": "#/definitions/uiContainer"
        }
      }
    },
    "selfServiceVerificationFlowState": {
      "description": "The state represents the state of the verification flow.\n\nchoose_method: ask the user to choose a method (e.g. recover account via email)\nsent_email: the email has been sent to the user\npassed_challenge: the request was successful and the recovery challenge was passed.",
      "type": "string",
      "title": "Verification Flow State"
    },
...
    "submitSelfServiceVerificationFlowBody": {
      "description": "nolint:deadcode,unused",
      "type": "object"
    },
...
    "uiContainer": {
      "description": "Container represents a HTML Form. The container can work with both HTTP Form and JSON requests",
      "type": "object",
      "required": [
        "action",
        "method",
        "nodes"
      ],
      "properties": {
        "action": {
          "description": "Action should be used as the form action URL `\u003cform action=\"{{ .Action }}\" method=\"post\"\u003e`.",
          "type": "string"
        },
        "messages": {
          "$ref": "#/definitions/uiTexts"
        },
        "method": {
          "description": "Method is the form method (e.g. POST)",
          "type": "string"
        },
        "nodes": {
          "$ref": "#/definitions/uiNodes"
        }
      }
    },
    "uiNode": {
      "description": "Nodes are represented as HTML elements or their native UI equivalents. For example,\na node can be an `\u003cimg\u003e` tag, or an `\u003cinput element\u003e` but also `some plain text`.",
      "type": "object",
      "title": "Node represents a flow's nodes",
      "required": [
        "type",
        "group",
        "attributes",
        "messages",
        "meta"
      ],
      "properties": {
        "attributes": {
          "$ref": "#/definitions/uiNodeAttributes"
        },
        "group": {
          "$ref": "#/definitions/uiNodeGroup"
        },
        "messages": {
          "$ref": "#/definitions/uiTexts"
        },
        "meta": {
          "$ref": "#/definitions/uiNodeMeta"
        },
        "type": {
          "$ref": "#/definitions/uiNodeType"
        }
      }
    },
...
    "uiNodeAttributes": {
      "type": "object",
      "title": "Attributes represents a list of attributes (e.g. `href=\"foo\"` for links)."
    },
    "uiNodeGroup": {
      "type": "string"
    },
...
    "uiNodeMeta": {
      "description": "This might include a label and other information that can optionally\nbe used to render UIs.",
      "type": "object",
      "title": "A Node's Meta Information",
      "properties": {
        "label": {
          "$ref": "#/definitions/uiText"
        }
      }
    },
...
    "uiNodeType": {
      "type": "string"
    },
    "uiNodes": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/uiNode"
      }
    },
    "uiText": {
      "type": "object",
      "required": [
        "id",
        "text",
        "type"
      ],
      "properties": {
        "context": {
          "description": "The message's context. Useful when customizing messages.",
          "type": "object"
        },
        "id": {
          "$ref": "#/definitions/ID"
        },
        "text": {
          "description": "The message text. Written in american english.",
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/uiTextType"
        }
      }
    },
    "uiTextType": {
      "type": "string"
    },
    "uiTexts": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/uiText"
      }
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is typically 201."
    }
  },
  "securityDefinitions": {
    "oryAccessToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}